fix: assorted fixes from pair
