+++ docs/examples/python/full/main.py
@@ -7,10 +7,10 @@ from glom import glom
 import sys
 
 sys.path.insert(1, '../../../clients/python')
-from aptible_client.helpers import getters, logger_utils, waiters  # NOQA
+from aptible_client.helpers import misc, logger_utils, waiters  # NOQA
 
 if typing.TYPE_CHECKING:
-    from clients.python.aptible_client.helpers import getters, logger_utils, waiters
+    from clients.python.aptible_client.helpers import misc, logger_utils, waiters
 
 logger = logger_utils.setup_logger("full")
 
@@ -37,7 +37,7 @@ def main(
         cleanup: Optional[bool] = False
 ):
     logger.info("Starting the full flow")
-    configuration = getters.get_client_configuration()
+    configuration = misc.get_client_configuration()
     # Initialize the API and create a session
     waiter = waiters.Waiter(
         configuration=configuration,
