+++ docs/examples/python/partials/vpc/main.py
@@ -5,10 +5,10 @@ from typing import Optional
 # below is done to enable testing outside the scope of the project with ease
 import sys
 sys.path.insert(1, '../../../clients/python')
-from aptible_client.helpers import getters, logger_utils, waiters  # NOQA
+from aptible_client.helpers import misc, logger_utils, waiters  # NOQA
 
 if typing.TYPE_CHECKING:
-    from clients.python.aptible_client.helpers import getters, logger_utils, waiters
+    from clients.python.aptible_client.helpers import misc, logger_utils, waiters
 
 
 logger = logger_utils.setup_logger("vpc")
@@ -27,7 +27,7 @@ def main(
     logger.info("Starting the create vpc flow")
     # Initialize the API and create a session
     waiter = waiters.Waiter(
-        configuration=getters.get_client_configuration(),
+        configuration=misc.get_client_configuration(),
         environment_id=environment_id,
         logger=logger,
         organization_id=organization_id,
